Fix printing of size_t.
git-svn-id: svn://svn.h5l.se/heimdal/trunk/heimdal@15942 ec53bebd-3082-4978-b11e-865c3cabbd6b

+++ lib/krb5/test_crypto_wrapping.c
@@ -73,8 +73,8 @@ test_wrapping(krb5_context context,
 
 	ret = krb5_encrypt(context, crypto, 0, buf, size, &data);
 	if (ret)
-	    krb5_err(context, 1, ret, "encrypt size %d using %s",
-		     size, etype_name);
+	    krb5_err(context, 1, ret, "encrypt size %lu using %s",
+		     (unsigned long)size, etype_name);
 
 	wrapped_size = krb5_get_wrapped_length(context, crypto, size);
